fix: force submodule to point at right upstream version
Our submodule was pointing at un unpublished rev. The reason is that we
have the submodule set to ignore changes (thanks to a2bd01439), and so
our `git add` command in `ck.sh` wasn't having any effect, and we ended
up committing a reference to something that we only had in one of our
local repos (probably a temporary "liferay" branch).

This commit adds an `-f` to the `git add` invocation to prevent this
from happening again, and forces the submodule to point at a published
commit.