chore: refresh built files
Just confirming that the parent commit really fixes things. I created a
fresh clone of the repo, ran `./ck.sh setup` (it worked), then `./ck.sh
build`, and this is the result. As you can see, only the
timestamp-related aspects have changed; we don't need to cut a new
release.