fix: make menubutton dependency explicit
This one is apparently required by scayt, but somehow, when we added
scayt, it actually got removed from the build output...

Whatever the case, I'm adding it back explicitly. Funnily enough,
re-doing the build with this in the list, we still don't see the
menubutton/plugin.js getting added back. Still not sure why that might
be. It's definitely included in the CKEditor source tree:

https://github.com/ckeditor/ckeditor4/tree/4.13.1/plugins/menubutton