fix: make dependencies implicit
This is absolutely crazy, but it seems that we don't get the
"menubutton" plugin unless we drop the dependencies. But if we drop the
dependencies, we still get the plugins!

Note the contents of the "scayt", "wsc" and "menubutton" directories in
the "ckeditor/plugins" folder.

I also tested with:

- Removing "wsc" from the list (files wind up in build anyway).
- Removing just "scayt" from the list (files wind up in build anyway).

So, it seems that these really are special dependencies, which work
magically when they are dropped into the right place, but which break
when you try to explicitly configure them.