feat: add Prettier
I want to streamline releases by using liferay-js-publish, so I may as
well follow the best practice recommended in the README there of adding
a "yarn ci" check to "preversion" script. We don't have any tests to run
as such, so we may as well check the format. (Later on, we can add tests
to confirm that the build output looks right etc).

https://github.com/liferay/liferay-npm-tools/tree/master/packages/liferay-js-publish