docs: add CHANGELOG
Created this initial one with a made-up-version number (v4.13.2),
because I'm not actually about to cut a release:

    npx liferay-changelog-generator \
      --regenerate \
      --from=4.11.3-liferay.0 \
      --version=v4.13.2

I then went in and edited the top (the fake release) and the bottom (the
very first version, which needed to be done manually because there were
no merge commits before the first version, and merge commits are what we
use to populate the changelog).

There is only one errant thing in the changelog, which is that the
"same" change -- "LPS-107869 Update CKEditor to latest version
(v4.13.1)" -- is featured in two versions. This is because we created
and merged two PRs with identical titles:

- https://github.com/liferay/liferay-ckeditor/pull/36
- https://github.com/liferay/liferay-ckeditor/pull/37

Rather than editing the changelog to distort what actually happened, we
may as well just let the underlying messiness show through.

Another gotcha here; we had a bad tag in the repo that caused the
changelog generator from running:

```
tag v4.11.liferay.2
Tagger: Julien Castelain <julien.castelain@liferay.com>
Date:   Tue May 28 15:02:28 2019 +0200

v4.11.4-liferay.2

commit 8da4564e687c54c4119c2f470105af619e4bd2c8 (tag: v4.11.4-liferay.2)
Author: Julien Castelain <julien.castelain@liferay.com>
Date:   Tue May 28 14:57:19 2019 +0200

    chore: Prepare for release 4.11.4-liferay.2
```

This was causing Git to freak out with:

```
warning: tag 'v4.11.liferay.2' is really 'v4.11.4-liferay.2' here
```

The fix was to delete the v4.11.4-liferay.2 tag and recreate it. Now it
looks like this:

```
tag v4.11.4-liferay.2
Tagger: Greg Hurrell <greg.hurrell@liferay.com>
Date:   Thu Feb 13 14:36:00 2020 +0100

v4.11.4-liferay.2

commit 8da4564e687c54c4119c2f470105af619e4bd2c8 (tag: v4.11.4-liferay.2)
Author: Julien Castelain <julien.castelain@liferay.com>
Date:   Tue May 28 14:57:19 2019 +0200

    chore: Prepare for release 4.11.4-liferay.2

```

With that change made (and pushed to GitHub), the changelog generator
worked. I expect subsequent updates should go smoothly, and because of
https://github.com/liferay/liferay-ckeditor/pull/41 we are unlikely to
have a bad tag created in the future.