chore: tweak Prettier in order to guarantee stable CHANGELOG
Comparing the effect of running `yarn format` in this repo with others,
it was changing the indent of bullet lists in our CHANGELOG in a way
that doesn't match the other repos. This is a problem because we have
tuned the output of liferay-changelog-generator to match what `prettier
format` expects — so we need it to behave exactly the same in all repos.

This commit adds the minimal config necessary to make the Markdown
output match, but note that because it is repo-wide, it also affects the
JS and JSON files, which seems fine.