feat: reduce noise created by patch updates
As described here:

https://github.com/liferay/liferay-ckeditor/pull/108#issuecomment-669938087

Some of the noise is due to the user doing the update using a different
version of Git, so the version number changes.

Some of it is due to the total number of patches changes, so each
subject must change too.

We can avoid both of these by adding the `--no-signature` and
`--no-numbered` switches to our `git format-patch` call.

Note though that we can't really stop the SHA-1's from changing if a
rebase or other mutation is involved, but they will probable stay stable
otherwise.